Accept float for usleep too and use one robust/universal sleep impl.

+++ ext/standard/basic_functions.stub.php
@@ -285,7 +285,7 @@ function flush(): void {}
 
 function sleep(float $seconds): int|float|false {}
 
-function usleep(int $microseconds): void {}
+function usleep(float $microseconds): void {}
 
 #if HAVE_NANOSLEEP
 function time_nanosleep(int $seconds, int $nanoseconds): array|bool {}
